requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pydantic>=2.9.0
pydantic-settings>=2.1.0
langchain>=0.1.0
langchain-core>=0.1.10

